requirements.txt
# Economic Analysis AI Agents with LangGraph
# Python 3.11+ required for LangGraph compatibility

# Core LangGraph and AI frameworks
langgraph>=0.2.16
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
yfinance>=0.2.18
fredapi>=0.5.0

# Visualization and charting
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.17.0

# Statistical analysis
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Data validation and utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Optional: Enhanced features
jupyter>=1.0.0  # For notebook analysis
streamlit>=1.28.0  # For web interface (optional)

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Performance optimization
numba>=0.58.0  # For numerical computations
joblib>=1.3.0  # For parallel processing